Group locks by recoverable status
