+++ packages/ui/src/app/pages/Profile/MyAccounts.tsx
@@ -1,28 +1,13 @@
-import React, { useState } from 'react'
+import React from 'react'
 
+import { Accounts } from '@/accounts/components/Accounts'
 import { useMyTotalBalances } from '@/accounts/hooks/useMyTotalBalances'
-import { RecoverBalanceModal } from '@/accounts/modals/RecoverBalance'
 import { PageLayout, PageHeaderWrapper } from '@/app/components/PageLayout'
-import { ButtonPrimary } from '@/common/components/buttons'
 import { RowGapBlock } from '@/common/components/page/PageContent'
 import { PageTitle } from '@/common/components/page/PageTitle'
 import { Statistics, TokenValueStat } from '@/common/components/statistics'
-import { BN_ZERO } from '@/common/constants'
 
-import { Accounts } from './components/Accounts'
 import { MyProfileTabs } from './components/MyProfileTabs'
-
-const RecoverBalances = () => {
-  const [isOpen, setOpen] = useState(false)
-  return (
-    <>
-      <ButtonPrimary size="medium" onClick={() => setOpen(true)}>
-        Recover all
-      </ButtonPrimary>
-      {isOpen && <RecoverBalanceModal onClose={() => setOpen(false)} />}
-    </>
-  )
-}
 
 export const MyAccounts = () => {
   const { total, transferable, locked, recoverable } = useMyTotalBalances()
@@ -39,9 +24,7 @@ export const MyAccounts = () => {
             <TokenValueStat title="Total balance" tooltipText="Lorem ipsum..." value={total} />
             <TokenValueStat title="Total transferable balance" tooltipText="Lorem ipsum..." value={transferable} />
             <TokenValueStat title="Total locked balance" tooltipText="Lorem ipsum..." value={locked} />
-            <TokenValueStat title="Total recoverable" tooltipText="Lorem ipsum..." value={recoverable}>
-              {recoverable.gt(BN_ZERO) ? <RecoverBalances /> : null}
-            </TokenValueStat>
+            <TokenValueStat title="Total recoverable" tooltipText="Lorem ipsum..." value={recoverable} />
           </Statistics>
         </RowGapBlock>
       }
