+++ packages/ui/test/accounts/components/MyAccounts.test.tsx
@@ -5,8 +5,8 @@ import BN from 'bn.js'
 import React from 'react'
 import { HashRouter } from 'react-router-dom'
 
+import { Accounts } from '@/accounts/components/Accounts'
 import { Account, Balances } from '@/accounts/types'
-import { Accounts } from '@/app/pages/Profile/components/Accounts'
 import { shortenAddress } from '@/common/model/formatters'
 import { KeyringContext } from '@/common/providers/keyring/context'
 import { MembershipContext } from '@/memberships/providers/membership/context'
